--- src/main/java/kr/ac/hs/selab/coreQa/dto/bundle/CoreQaCreateBundle.java
@@ -1,4 +1,0 @@
-package kr.ac.hs.selab.coreQa.dto.bundle;
-
-public record CoreQaCreateBundle(String email, String title, String content) {
-}
