--- src/main/java/kr/ac/hs/selab/coreQa/dto/response/CoreQaCreateResponse.java
@@ -1,4 +1,0 @@
-package kr.ac.hs.selab.coreQa.dto.response;
-
-public record CoreQaCreateResponse(String title, String content) {
-}
