--- src/main/java/kr/ac/hs/selab/coreQa/dto/response/CoreQaReadResponse.java
@@ -1,4 +1,0 @@
-package kr.ac.hs.selab.coreQa.dto.response;
-
-public record CoreQaReadResponse(String title, String content) {
-}
