+++ src/main/java/kr/ac/hs/selab/core_qa/dto/bundle/CoreQaCreateBundle.java
@@ -1,0 +1,4 @@
+package kr.ac.hs.selab.core_qa.dto.bundle;
+
+public record CoreQaCreateBundle(String email, String title, String content) {
+}
