+++ src/main/java/kr/ac/hs/selab/core_qa/dto/response/CoreQaCreateResponse.java
@@ -1,0 +1,4 @@
+package kr.ac.hs.selab.core_qa.dto.response;
+
+public record CoreQaCreateResponse(String title, String content) {
+}
