+++ src/main/java/kr/ac/hs/selab/core_qa/dto/response/CoreQaReadResponse.java
@@ -1,0 +1,4 @@
+package kr.ac.hs.selab.core_qa.dto.response;
+
+public record CoreQaReadResponse(String title, String content) {
+}
