--- src/main/java/kr/ac/hs/selab/postLike/dto/PostLikeFindDto.java
@@ -1,10 +1,0 @@
-package kr.ac.hs.selab.postLike.dto;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public class PostLikeFindDto {
-    private final Long postId;
-}
