--- src/main/java/kr/ac/hs/selab/postLike/dto/response/PostLikeResponse.java
@@ -1,10 +1,0 @@
-package kr.ac.hs.selab.postLike.dto.response;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public class PostLikeResponse {
-    private final Long id;
-}
