+++ src/main/java/kr/ac/hs/selab/post_like/dto/PostLikeFindDto.java
@@ -1,0 +1,10 @@
+package kr.ac.hs.selab.post_like.dto;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public class PostLikeFindDto {
+    private final Long postId;
+}
