+++ src/main/java/kr/ac/hs/selab/post_like/dto/response/PostLikeResponse.java
@@ -1,0 +1,10 @@
+package kr.ac.hs.selab.post_like.dto.response;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public class PostLikeResponse {
+    private final Long id;
+}
