+++ src/main/java/kr/ac/hs/selab/post_like/infrastructure/PostLikeRepository.java
@@ -1,0 +1,10 @@
+package kr.ac.hs.selab.post_like.infrastructure;
+
+import kr.ac.hs.selab.post_like.domain.PostLike;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
+    List<PostLike> findByPostId(Long postId);
+}
